pyproject.toml
[project]
name = "english-tutor"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.3",
    "langchain-google-genai>=4.1.3",
    "plotly>=6.5.1",
    "psycopg-pool>=3.3.0",
    "psycopg[binary]>=3.3.2",
    "py-pglite>=0.5.3",
    "pyyaml>=6.0.3",
    "reflex>=0.8.24.post1",
    "ty>=0.0.11",
]
